requirements.txt
attrs==25.3.0
blinker==1.9.0
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cmake==4.1.0
colorzero==2.0
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
evdev==1.9.2
expecttest==0.3.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.0
fsspec==2025.9.0
future==1.0.0
gpiozero==2.0.1
hypothesis==6.140.2
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lgpio==0.2.2.0
lintrunner==0.12.7
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
optree==0.17.0
packaging==25.0
pigpio==1.78
pillow==11.3.0
polars==1.33.1
psutil==7.1.0
pynput==1.8.1
pyparsing==3.2.5
pyproject_hooks==1.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-vlc==3.0.21203
python-xlib==0.33
pytube==15.0.0
PyYAML==6.0.2
requests==2.32.5
RPi.GPIO==0.7.1
scipy==1.16.2
serial==0.0.97
six==1.17.0
sortedcontainers==2.4.0
sympy==1.14.0
tkinterweb==4.4.4
tkinterweb-tkhtml==1.1.4
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
uv==0.8.22
Werkzeug==3.1.3
